TVGUI-17 Move TVGuide component into components directory
