+++ src/Pages/ProfilePage.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect, useContext } from "react";
 import { useParams } from "react-router";
 import { Alert, Button, Form, List, Modal, Select, Space } from "antd";
 
-import TVGuide from "../TVGuide/TVGuide";
+import TVGuide from "../components/TVGuide";
 import { UserContext } from "../contexts/UserContext";
 import { getGuide, getRecordedShows, getReminders, getUser, updateSubscriptions } from "../requests/requests";
 import { Guide, RecordedShowModel, Reminder, SubscriptionsPayload, User } from "../utils";
